fix create product

+++ se1979-su25-g2/se1979-su-25-g2-be/src/main/java/org/example/se1979su25g2be/service/OrderServiceImpl.java
@@ -43,7 +43,7 @@ public class OrderServiceImpl implements OrderService {
         User user = null;
 
         if (orderRequest.getUserId() != null) {
-            user = userRepository.findById(orderRequest.getUserId().longValue())
+            user = userRepository.findById((int) orderRequest.getUserId().longValue())
                     .orElseThrow(() -> new EntityNotFoundException("User not found"));
             cartItems = cartItemRepository.findByUser(user);
         } else if (orderRequest.getSessionId() != null && !orderRequest.getSessionId().isEmpty()) {
